Share elasticsearch configuration
