+++ src/main/java/ca/uhn/fhir/jpa/starter/FhirServerConfigDstu3.java
@@ -3,7 +3,6 @@ package ca.uhn.fhir.jpa.starter;
 import ca.uhn.fhir.context.ConfigurationException;
 import ca.uhn.fhir.jpa.config.BaseJavaConfigDstu3;
 import ca.uhn.fhir.jpa.search.DatabaseBackedPagingProvider;
-import ca.uhn.fhir.jpa.search.lastn.ElasticsearchSvcImpl;
 import ca.uhn.fhir.jpa.starter.annotations.OnDSTU3Condition;
 import ca.uhn.fhir.jpa.starter.cql.StarterCqlDstu3Config;
 import javax.annotation.PostConstruct;
@@ -22,7 +21,7 @@ import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 
 @Configuration
 @Conditional(OnDSTU3Condition.class)
-@Import(StarterCqlDstu3Config.class)
+@Import({StarterCqlDstu3Config.class, ElasticsearchConfig.class})
 public class FhirServerConfigDstu3 extends BaseJavaConfigDstu3 {
 
   @Autowired
@@ -87,20 +86,4 @@ public class FhirServerConfigDstu3 extends BaseJavaConfigDstu3 {
     retVal.setEntityManagerFactory(entityManagerFactory);
     return retVal;
   }
-
-  @Bean()
-  public ElasticsearchSvcImpl elasticsearchSvc() {
-    if (EnvironmentHelper.isElasticsearchEnabled(configurableEnvironment)) {
-      String elasticsearchUrl = EnvironmentHelper.getElasticsearchServerUrl(configurableEnvironment);
-		 if (elasticsearchUrl.startsWith("http")) {
-			 elasticsearchUrl =elasticsearchUrl.substring(elasticsearchUrl.indexOf("://") + 3);
-		 }
-      String elasticsearchUsername = EnvironmentHelper.getElasticsearchServerUsername(configurableEnvironment);
-      String elasticsearchPassword = EnvironmentHelper.getElasticsearchServerPassword(configurableEnvironment);
-      return new ElasticsearchSvcImpl(elasticsearchUrl, elasticsearchUsername, elasticsearchPassword);
-    } else {
-      return null;
-    }
-  }
-
 }
